adding the beginning files to the project
added a ts folder to house the index.ts file, as well as the node modules and package.json files that come with any npm package. Furthermore, make a db folder for the database the currently holds a index.sql file. As time continues with this project, I am sure this file structure will change drastically, but here is a beginning place.