Update from Review Session
In this commit (which is in the main branch and is not supposed to be) are the updates that were completed during class review session on August 27th, 2024. This includes file structure, Beginning the schema.sql file by creating the tables within our database, beginning the seeds.sql file by inserting the stock values into the said tables, beginning the index.ts file with the first inquirer prompt that will repeat the options to the user.